Added support for VoxelGI placeholder materials
VoxelGI in the form it's currently available in doesn't support
reading the properties of ShaderMaterials and assumes an entirely
white albedo for them, resulting in incorrect (overexposed) GI.

This commits adds the voxelgi_placeholder property to
ShaderMaterial that can be used to link in a BaseMaterial3D. When
facing a ShaderMaterial, VoxelGI will now check if it has a
placeholder material set, and if it is, then it will pick the
relevant properties (albedo and emissive colors) from it. With no
placeholder material set, the old VoxelGI behavior of assuming
white albedo will remain in effect.