Added support for overriding animation loop mode.
This adds support for specifying a looping mode for an animation
at the playback time, possibly different from the one specified in
the animation resource.

The following features were added:

  - loop_override in Animation (try_)interpolate_* methods, for
    proper looping interpolation.

  - loop_override in AnimationPlayer play and play_backward methods.

  - loop_override property in AnimationNodeAnimation.

In all added methods, the loop_override parameter was added at the
end, with a default value signifying 'do not override', so the API
compatibility was kept for existing users.